crawler/requirements.txt
# Baoyan crawler service - Python dependencies
# Install: pip install -r requirements.txt

# Crawler framework
Scrapy==2.11.0

# Async HTTP
aiohttp==3.9.0
aiofiles==23.2.0

# HTML parsing
lxml==4.9.3
parsel==1.8.1
beautifulsoup4==4.12.2
pypdf==4.3.1

# Database
PyMySQL==1.1.0
SQLAlchemy==2.0.23

# Redis
redis==5.0.1

# Scheduler
APScheduler==3.10.4

# HTTP client
requests==2.31.0

# Config
python-dotenv==1.0.0
PyYAML==6.0.1

# Logging
loguru==0.7.2

# Utilities
python-dateutil==2.8.2
pydantic==2.5.0

# Test
pytest==7.4.3
pytest-asyncio==0.21.1
